// src/features/todo/p0019/types.ts
export interface Profile {
  id: string; // should match user.id
  created_at?: string;
  updated_at: string | null;
  last_sign_in_at?: string | null;
  full_name: string | null;
  avatar_url: string | null;
  email?: string | null; // Added email field
  role: 'admin' | 'manager' | 'employee';
  default_project_id?: number | null;
}
// ...
export interface Project {
  id: number;
  created_at: string;
  name: string;
  created_by: string;
  project_members?: { count: number }[];
  color?: string | null;
}

export interface ProjectMember {
    project_id: number;
    user_id: string;
    created_at: string;
    projects: Project | null;
    profiles?: Profile;
}

export interface MemberDetails extends ProjectMember {
    profiles: Profile;
    task_count: number;
}